[fix] MFINDBUGS-145 needs site plugin 3.6 but needs doxia upgraded
For jdk 14 and above, doxia needs upgraded to work.  This seemed to come and go as jdk 14 was going through updates.  It is now in rampdown and latest openjdk which is listed as oracle and probably close to last build now fails again and can be confirmed on both linux and windows.  While jdk 15 still passes, it is more likely due to that not being jdk 15 really other than adjustments to overall product to indicate such.  In otherwords, it is unlikley to have the hard requirements of jdk 14 still for the most part.