[windows] Add fix to support processing properly on windows by not ending up with long c_ path to jar files
Fix on #238 only resulted in cleaning up unnecessary copies but made other processing worse.  As a result, #238 is being reverted.  This patch is to make windows usage behave like linux usage.  It still expected to copy the jar files (even if unnecessary) but not nexted into a deep folder structure, rather just the artifact name.  The original logic did not handle windows at all.  It now does but is still not a great solution.