[site] Fix site on github actions
Not sure exactly how it gets the folder name other than that is part of scm piece site was moved onto for deployment.  This sort of made sense it has use of a super pom and the site gets messed up trying to push back to a parent.  I would have thought name would work out differently then